+++ R/batchMap.R
@@ -78,9 +78,9 @@ batchMap = function(fun, ..., args = list(), more.args = list(), reg = getDefaul
     return(noIds())
   info("Adding %i jobs ...", nrow(ddd))
 
-  writeRDS(fun, file = file.path(reg$file.dir, "user.function.rds"))
+  writeRDS(fun, file = fp(reg$file.dir, "user.function.rds"))
   if (length(more.args) > 0L)
-    writeRDS(more.args, file = file.path(reg$file.dir, "more.args.rds"))
+    writeRDS(more.args, file = fp(reg$file.dir, "more.args.rds"))
   ids = seq_row(ddd)
 
   reg$defs = data.table(
@@ -100,6 +100,7 @@ batchMap = function(fun, ..., args = list(), more.args = list(), reg = getDefaul
     batch.id    = NA_character_,
     log.file    = NA_character_,
     job.hash    = NA_character_,
+    job.name    = NA_character_,
     key         = "job.id")
 
   saveRegistry(reg)