+++ R/findJobs.R
@@ -2,11 +2,13 @@
 #'
 #' @description
 #' These functions are used to find and filter jobs, depending on either their parameters (\code{findJobs} and
-#' \code{findExperiments}), their tags (\code{findTagged}), or their computational status (all other functions).
+#' \code{findExperiments}), their tags (\code{findTagged}), or their computational status (all other functions,
+#' see \code{\link{getStatus}} for an overview).
 #'
-#' For a summarizing overview over the status, see \code{\link{getStatus}}.
-#' Note that \code{findOnSystem} and \code{findExpired} are somewhat heuristic and may report misleading results,
-#' depending on the state of the system and the \code{\link{ClusterFunctions}} implementation.
+#' Note that \code{findQueued}, \code{findRunning}, \code{findOnSystem} and \code{findExpired} are somewhat heuristic
+#' and may report misleading results, depending on the state of the system and the \code{\link{ClusterFunctions}} implementation.
+#'
+#' See \code{\link{JoinTables}} for convenient set operations (unions, intersects, differences) on tables with job ids.
 #'
 #' @param expr [\code{expression}]\cr
 #'   Predicate expression evaluated in the job parameters.
@@ -15,6 +17,7 @@
 #' @template ids
 #' @template reg
 #' @return [\code{\link{data.table}}] with column \dQuote{job.id} containing matched jobs.
+#' @seealso \code{\link{getStatus}} \code{\link{JoinTables}}<`3`>
 #' @export
 #' @examples
 #' tmp = makeRegistry(file.dir = NA, make.default = FALSE)
@@ -217,6 +220,13 @@ findErrors = function(ids = NULL, reg = getDefaultRegistry()) {
 }
 
 
+# used in waitForJobs: find jobs which are done or error
+.findTerminated = function(reg, ids = NULL) {
+  done = NULL
+  filter(reg$status, ids, c("job.id", "done"))[!is.na(done), "job.id"]
+}
+
+
 #' @export
 #' @rdname findJobs
 findOnSystem = function(ids = NULL, reg = getDefaultRegistry()) {
@@ -255,7 +265,7 @@ findExpired = function(ids = NULL, reg = getDefaultRegistry()) {
 
 .findExpired = function(reg, ids = NULL, batch.ids = getBatchIds(reg)) {
   submitted = done = batch.id = NULL
-  filter(reg$status, ids, c("job.id", "submitted", "done", "batch.id"))[!is.na(submitted) & is.na(done) & batch.id %nin% batch.ids$batch.id, "job.id"]
+  filter(reg$status, ids, c("job.id", "submitted", "done", "batch.id"))[!is.na(submitted) & is.na(done) & batch.id %chnin% batch.ids$batch.id, "job.id"]
 }
 
 #' @export