+++ R/updateRegisty.R
@@ -1,8 +1,9 @@
+# returns TRUE if the state possibly changed
 updateRegistry = function(reg = getDefaultRegistry()) { # nocov start
   "!DEBUG [updateRegistry]: Running updateRegistry"
   pv = packageVersion("batchtools")
   if (identical(pv, reg$version))
-    return(TRUE)
+    return(FALSE)
 
   if (is.null(reg$version) || reg$version < "0.9.0")
     stop("Your registry is too old.")
@@ -16,19 +17,20 @@ updateRegistry = function(reg = getDefaultRegistry()) { # nocov start
     }
 
     ### hotfix for log.file column
-    if ("log.file" %nin% names(reg$status)) {
-      reg$status$log.file = NA_character_
+    if ("log.file" %chnin% names(reg$status)) {
+      info("Adding column 'log.file'")
+      reg$status[, ("log.file") := rep(NA_character_, .N)]
     }
   }
 
   if (reg$version < "0.9.1-9001") {
     ### hotfix for base32 encoding of exports
-    fns = list.files(file.path(reg$file.dir, "exports"), pattern = "\\.rds$", all.files = TRUE, no.. = TRUE)
+    fns = list.files(fp(reg$file.dir, "exports"), pattern = "\\.rds$", all.files = TRUE, no.. = TRUE)
     if (length(fns)) {
       info("Renaming export files")
       file.rename(
-        file.path(reg$file.dir, fns),
-        file.path(reg$file.dir, mangle(stri_sub(fns, to = -5L)))
+        fp(reg$file.dir, fns),
+        fp(reg$file.dir, mangle(stri_sub(fns, to = -5L)))
       )
     }
   }
@@ -36,11 +38,18 @@ updateRegistry = function(reg = getDefaultRegistry()) { # nocov start
   if (reg$version < "0.9.1-9002" && inherits(reg, "ExperimentRegistry")) {
     info("Renaming problems and algorithm files")
     for (prob in getProblemIds(reg))
-      file.rename(file.path(reg$file.dir, "problems", sprintf("%s.rds", digest(prob))), getProblemURI(reg, prob))
+      file.rename(fp(reg$file.dir, "problems", sprintf("%s.rds", digest(prob))), getProblemURI(reg, prob))
     for (algo in getAlgorithmIds(reg))
-      file.rename(file.path(reg$file.dir, "algorithms", sprintf("%s.rds", digest(algo))), getAlgorithmURI(reg, algo))
+      file.rename(fp(reg$file.dir, "algorithms", sprintf("%s.rds", digest(algo))), getAlgorithmURI(reg, algo))
+  }
+
+  if (reg$version < "0.9.4-9001") {
+    if ("job.name" %chnin% names(reg$status)) {
+      info("Adding column 'job.name'")
+      reg$status[, ("job.name") := rep(NA_character_, .N)]
+    }
   }
 
   reg$version = pv
-  saveRegistry(reg)
+  return(TRUE)
 } # nocov end