+++ R/waitForJobs.R
@@ -6,13 +6,25 @@
 #' @templateVar ids.default findSubmitted
 #' @template ids
 #' @param sleep [\code{function(i)} | \code{numeric(1)}]\cr
-#'   Function which returns the duration to sleep in the \code{i}-th iteration.
-#'   Alternatively, you can pass a single positive numeric value.
+#'   Parameter to control the duration to sleep between queries.
+#'   You can pass an absolute numeric value in seconds or a \code{function(i)} which returns
+#'   the number of seconds to sleep in the \code{i}-th iteration.
+#'   If not provided (\code{NULL}), tries to read the value (number/function) from the configuration file
+#'   (stored in \code{reg$sleep}) or defaults to a function with exponential backoff between
+#'   5 and 120 seconds.
 #' @param timeout [\code{numeric(1)}]\cr
 #'   After waiting \code{timeout} seconds, show a message and return
 #'   \code{FALSE}. This argument may be required on some systems where, e.g.,
 #'   expired jobs or jobs on hold are problematic to detect. If you don't want
 #'   a timeout, set this to \code{Inf}. Default is \code{604800} (one week).
+#' @param expire.after [\code{integer(1)}]\cr
+#'   Jobs count as \dQuote{expired} if they are not found on the system but have not communicated back
+#'   their results (or error message). This frequently happens on managed system if the scheduler kills
+#'   a job because the job has hit the walltime or request more memory than reserved.
+#'   On the other hand, network file systems often require several seconds for new files to be found,
+#'   which can lead to false positives in the detection heuristic.
+#'   \code{waitForJobs} treats such jobs as expired after they have not been detected on the system
+#'   for \code{expire.after} iterations (default 3 iterations).
 #' @param stop.on.error [\code{logical(1)}]\cr
 #'   Immediately cancel if a job terminates with an error? Default is
 #'   \code{FALSE}.
@@ -21,83 +33,75 @@
 #'   successfully and \code{FALSE} if either the timeout is reached or at least
 #'   one job terminated with an exception.
 #' @export
-waitForJobs = function(ids = NULL, sleep = default.sleep, timeout = 604800, stop.on.error = FALSE, reg = getDefaultRegistry()) {
+waitForJobs = function(ids = NULL, sleep = NULL, timeout = 604800, expire.after = 3L, stop.on.error = FALSE, reg = getDefaultRegistry()) {
   assertRegistry(reg, writeable = FALSE, sync = TRUE)
   assertNumber(timeout, lower = 0)
+  assertCount(expire.after, positive = TRUE)
   assertFlag(stop.on.error)
-  sleep = getSleepFunction(sleep)
+  sleep = getSleepFunction(reg, sleep)
   ids = convertIds(reg, ids, default = .findSubmitted(reg = reg))
-
-  .findNotTerminated = function(reg, ids = NULL) {
-    done = NULL
-    filter(reg$status, ids, c("job.id", "done"))[is.na(done), "job.id"]
-  }
 
   if (nrow(.findNotSubmitted(ids = ids, reg = reg)) > 0L) {
     warning("Cannot wait for unsubmitted jobs. Removing from ids.")
     ids = ids[.findSubmitted(ids = ids, reg = reg), nomatch = 0L]
   }
 
-  n.jobs = nrow(ids)
-  if (n.jobs == 0L)
+  if (nrow(ids) == 0L) {
     return(TRUE)
+  }
 
-  batch.ids = getBatchIds(reg)
-  "!DEBUG [waitForJobs]: Using `nrow(ids)` ids and `nrow(batch.ids)` initial batch ids"
+  terminated = on.sys = expire.counter = NULL
+  ids$terminated = FALSE
+  ids$on.sys = FALSE
+  ids$expire.counter = 0L
 
   timeout = Sys.time() + timeout
-  ids.disappeared = noIds()
-  pb = makeProgressBar(total = n.jobs, format = "Waiting (S::system R::running D::done E::error) [:bar] :percent eta: :eta")
-  i = 1L
+  pb = makeProgressBar(total = nrow(ids), format = "Waiting (S::system R::running D::done E::error) [:bar] :percent eta: :eta")
+  i = 0L
 
   repeat {
-    # case 1: all jobs terminated -> nothing on system
-    ids.nt = .findNotTerminated(reg, ids)
-    if (nrow(ids.nt) == 0L) {
+    ### case 1: all jobs terminated -> nothing on system
+    ids[.findTerminated(reg, ids), "terminated" := TRUE]
+    if (ids[!(terminated), .N] == 0L) {
       "!DEBUG [waitForJobs]: All jobs terminated"
       pb$update(1)
       waitForResults(reg, ids)
       return(nrow(.findErrors(reg, ids)) == 0L)
     }
 
-    stats = getStatusTable(ids = ids, batch.ids = batch.ids, reg = reg)
-    pb$update((n.jobs - nrow(ids.nt)) / n.jobs, tokens = as.list(stats))
-
-    # case 2: there are errors and stop.on.error is TRUE
+    ### case 2: there are errors and stop.on.error is TRUE
     if (stop.on.error && nrow(.findErrors(reg, ids)) > 0L) {
       "!DEBUG [waitForJobs]: Errors found and stop.on.error is TRUE"
       pb$update(1)
       return(FALSE)
     }
 
-    # case 3: we have reached a timeout
+    batch.ids = getBatchIds(reg)
+    ids[, "on.sys" := FALSE][.findOnSystem(reg, ids, batch.ids = batch.ids), "on.sys" := TRUE]
+    ids[!(on.sys) & !(terminated), "expire.counter" := expire.counter + 1L]
+    stats = getStatusTable(ids = ids, batch.ids = batch.ids, reg = reg)
+    pb$update(mean(ids$terminated), tokens = as.list(stats))
+    "!DEBUG [waitForJobs]: batch.ids: `stri_flatten(batch.ids$batch.id, ',')`"
+
+    ### case 3: jobs disappeared, we cannot find them on the system in [expire.after] iterations
+    if (ids[!(terminated) & expire.counter > expire.after, .N] > 0L) {
+      warning("Some jobs disappeared from the system")
+      pb$update(1)
+      waitForResults(reg, ids)
+      return(FALSE)
+    }
+
+    # case 4: we reach a timeout
+    sleep(i)
+    i = i + 1L
     if (Sys.time() > timeout) {
       pb$update(1)
       warning("Timeout reached")
       return(FALSE)
     }
 
-    # case 4: jobs disappeared, we cannot find them on the system
-    # heuristic:
-    #   job is not terminated, not on system and has not been on the system
-    #   in the previous iteration
-    ids.on.sys = .findOnSystem(reg, ids, batch.ids = batch.ids)
-    if (nrow(ids.disappeared) > 0L) {
-      if (nrow(ids.nt[!ids.on.sys, on = "job.id"][ids.disappeared, on = "job.id", nomatch = 0L]) > 0L) {
-        warning("Some jobs disappeared from the system")
-        pb$update(1)
-        waitForResults(reg, ids)
-        return(FALSE)
-      }
-    }
-
-    ids.disappeared = ids[!ids.on.sys, on = "job.id"]
-    "!DEBUG [waitForJobs]: `nrow(ids.disappeared)` jobs disappeared"
-
-    sleep(i)
-    i = 1 + 1L
-    suppressMessages(syncRegistry(reg = reg))
-    batch.ids = getBatchIds(reg)
-    "!DEBUG [waitForJobs]: New batch.ids: `stri_flatten(batch.ids$batch.id, ',')`"
+    if (suppressMessages(sync(reg = reg)))
+      saveRegistry(reg)
   }
 }
+