+++ tests/testthat/test_ClusterFunctions.R
@@ -4,7 +4,7 @@ test_that("clusterFunctions constructor", {
   check = function(cf) {
     expect_is(cf, "ClusterFunctions")
     expect_set_equal(names(cf), c("name", "submitJob", "killJob", "listJobsQueued", "listJobsRunning",
-        "store.job", "array.var", "scheduler.latency", "fs.latency", "hooks"))
+        "store.job.collection", "store.job.files", "array.var", "scheduler.latency", "fs.latency", "hooks"))
     expect_output(print(cf), "ClusterFunctions for mode")
   }
   reg = makeRegistry(file.dir = NA, make.default = FALSE)
@@ -76,7 +76,7 @@ test_that("Special chars in directory names", {
   base.dir = tempfile(pattern = "test", tmpdir = dirname(reg$file.dir))
   dir.create(base.dir, recursive = TRUE)
 
-  file.dir = file.path(base.dir, "test#some_frequently-used chars")
+  file.dir = fp(base.dir, "test#some_frequently-used chars")
   reg = makeRegistry(file.dir, make.default = FALSE)
   batchMap(identity, 1:2, reg = reg)
   submitAndWait(reg = reg)
@@ -90,9 +90,7 @@ test_that("Export of environment variable DEBUGME", {
     skip("Environment variables not exported for CF socket")
   batchMap(function(i) Sys.getenv("DEBUGME"), i = 1, reg = reg)
 
-  prev = Sys.getenv("DEBUGME")
-  on.exit(Sys.setenv(DEBUGME = prev))
-  Sys.setenv(DEBUGME = "grepme")
+  withr::local_envvar(c("DEBUGME" = "grepme"))
   submitAndWait(reg, 1)
 
   res = loadResult(1, reg = reg)