+++ tests/testthat/test_batchMap.R
@@ -6,8 +6,8 @@ test_that("batchMap", {
   ids = batchMap(fun, i = 1:3, more.args = list(x = 1), reg = reg)
   expect_data_table(ids, any.missing = FALSE, ncols = 1L, nrow = 3L, key = "job.id")
   expect_equal(ids$job.id, 1:3)
-  expect_equal(readRDS(file.path(reg$file.dir, "user.function.rds")), fun)
-  expect_equal(readRDS(file.path(reg$file.dir, "more.args.rds")), list(x = 1))
+  expect_equal(readRDS(fp(reg$file.dir, "user.function.rds")), fun)
+  expect_equal(readRDS(fp(reg$file.dir, "more.args.rds")), list(x = 1))
 
   checkTables(reg)
   expect_data_table(reg$defs, nrow = 3L, any.missing = FALSE)
@@ -18,13 +18,13 @@ test_that("batchMap", {
   reg = makeRegistry(file.dir = NA, make.default = FALSE)
   fun = function(...) list(...)
   ids = batchMap(fun, 1:3, more.args = list(j = 1), reg = reg)
-  expect_equal(readRDS(file.path(reg$file.dir, "more.args.rds")), list(j = 1))
+  expect_equal(readRDS(fp(reg$file.dir, "more.args.rds")), list(j = 1))
   expect_equal(reg$defs$pars, lapply(1:3, list))
 
   reg = makeRegistry(file.dir = NA, make.default = FALSE)
   fun = function(...) list(...)
   ids = batchMap(fun, args = list(1:3), more.args = list(j = 1), reg = reg)
-  expect_equal(readRDS(file.path(reg$file.dir, "more.args.rds")), list(j = 1))
+  expect_equal(readRDS(fp(reg$file.dir, "more.args.rds")), list(j = 1))
   expect_equal(reg$defs$pars, lapply(1:3, list))
 
   reg = makeRegistry(file.dir = NA, make.default = FALSE)