+++ tests/testthat/test_getJobTable.R
@@ -6,12 +6,12 @@ test_that("getJobTable.Registry", {
   ids = batchMap(fun, i = 1:4, j = rep(1, 4), reg = reg)
 
   tab = getJobTable(reg = reg, flatten = FALSE)
-  expect_data_table(tab, nrows = 4, ncols = 14, key = "job.id")
+  expect_data_table(tab, nrows = 4, ncols = 15, key = "job.id")
   expect_list(tab$pars)
   expect_equal(tab$pars[[1]], list(i = 1L, j = 1))
 
   tab = getJobTable(reg = reg, flatten = TRUE)
-  expect_data_table(tab, nrows = 4, ncols = 14, key = "job.id")
+  expect_data_table(tab, nrows = 4, ncols = 15, key = "job.id")
   expect_null(tab[["pars"]])
   expect_equal(tab$i, 1:4)
   expect_equal(tab$j, rep(1, 4))
@@ -119,7 +119,7 @@ test_that("getJobTable.ExperimentRegistry", {
   ids = addExperiments(list(p1 = data.table(k = 1)), list(a1 = data.table(sq = 1:3)), reg = reg)
 
   tab = getJobTable(reg = reg, flatten = FALSE)
-  expect_data_table(tab, nrows = 3, ncols = 17, key = "job.id")
+  expect_data_table(tab, nrows = 3, ncols = 18, key = "job.id")
   expect_copied(tab, reg$status)
   expect_list(tab$pars)
   expect_equal(tab$pars[[1]], list(prob.pars = list(k = 1), algo.pars = list(sq = 1)))
@@ -129,7 +129,7 @@ test_that("getJobTable.ExperimentRegistry", {
   expect_equal(tab$algorithm[1], factor("a1"))
 
   tab = getJobTable(ids = 1:3, reg = reg, flatten = TRUE)
-  expect_data_table(tab, nrows = 3, ncols = 17, key = "job.id")
+  expect_data_table(tab, nrows = 3, ncols = 18, key = "job.id")
   expect_null(tab[["pars"]])
   expect_set_equal(tab$k, rep(1, 3))
   expect_set_equal(tab$sq, 1:3)
@@ -148,4 +148,23 @@ test_that("flatten autodetection", {
   tab = getJobPars(reg = reg, flatten = NULL)
   expect_data_table(tab, ncols = 2, key = "job.id")
   expect_set_equal(names(tab), c("job.id", "pars"))
+})
+
+test_that("flatten heuristic for experiment registries (#123)", {
+  tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE)
+
+  fun = function(job, data, n, mean, sd, ...) rnorm(sum(n), mean = mean, sd = sd)
+  addProblem("rnorm", fun = fun, reg = tmp)
+  fun = function(instance, ...) sd(instance)
+  addAlgorithm("deviation", fun = fun, reg = tmp)
+
+  prob.designs = algo.designs = list()
+  prob.designs$rnorm = data.table(expand.grid(n = list(100, 1:4), mean = 0, sd = 1:2))
+  algo.designs$deviation = data.table()
+  addExperiments(prob.designs, algo.designs, reg = tmp)
+  submitAndWait(reg = tmp)
+
+  res = getJobPars(reg = tmp)
+  expect_data_table(res, ncol = 4)
+  expect_list(res$pars, len = 4)
 })