+++ tests/testthat/test_resetJobs.R
@@ -13,14 +13,14 @@ test_that("resetJobs", {
   expect_true(file.exists(getLogFiles(reg, 3)))
   expect_false(identical(reg$status$submitted, before$status$submitted))
   expect_true(file.exists(getResultFiles(reg, 1)))
-  expect_equal(dir.exists(file.path(reg$file.dir, "external", 1:3)), c(TRUE, FALSE, TRUE))
+  expect_equal(dir.exists(fp(reg$file.dir, "external", 1:3)), c(TRUE, FALSE, TRUE))
 
   resetJobs(1, reg = reg)
   expect_true(all.equal(before$status[1], reg$status[1]))
   expect_false(file.exists(getResultFiles(reg, 1)))
   expect_true(file.exists(getResultFiles(reg, 3)))
   expect_true(file.exists(getLogFiles(reg, 3)))
-  expect_equal(dir.exists(file.path(reg$file.dir, "external", 1:3)), c(FALSE, FALSE, TRUE))
+  expect_equal(dir.exists(fp(reg$file.dir, "external", 1:3)), c(FALSE, FALSE, TRUE))
   expect_false(file.exists(getResultFiles(reg, 1)))
   expect_true(file.exists(getResultFiles(reg, 3)))
 
@@ -30,7 +30,7 @@ test_that("resetJobs", {
   expect_equivalent(before$status, reg$status)
   expect_false(file.exists(getLogFiles(reg, 3)))
   expect_false(file.exists(getResultFiles(reg, 3)))
-  expect_equal(dir.exists(file.path(reg$file.dir, "external", 1:3)), c(FALSE, FALSE, FALSE))
+  expect_equal(dir.exists(fp(reg$file.dir, "external", 1:3)), c(FALSE, FALSE, FALSE))
 })